move espresso logic to arbos, handle nil justifications
